Fix redirect back specs for Rails 5.2
The mechanism for redirecting back changed between Rails 5.0 and 5.2;
in Rails 5.0, we could write `redirect_to :back` and risk an exception
being raised, but in Rails 5.2, we now use `redirect_back` and must
explicitly provide a fallback location. This is semantically different
enough that it can't be adopted directly here; we want a custom
redirection but the old Rails exception no longer exists.